rename: detail 패키지명 변경
- 이전: detail
- 이후: reservation
