+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:parentActivityName=".view.home.ReservationHomeActivity"
             android:exported="false" />
         <activity
-            android:name=".view.detail.ReservationDetailActivity"
+            android:name=".view.reservation.ReservationDetailActivity"
             android:parentActivityName=".view.home.ReservationHomeActivity"
             android:exported="false" />
     </application>
