+++ app/src/main/java/woowacourse/movie/view/finished/ReservationFinishedActivity.kt
@@ -12,9 +12,9 @@ import woowacourse.movie.model.Movie
 import woowacourse.movie.model.Ticket
 import woowacourse.movie.presenter.finished.ReservationFinishedContract
 import woowacourse.movie.presenter.finished.ReservationFinishedPresenter
-import woowacourse.movie.view.detail.ReservationDetailActivity.Companion.DEFAULT_MOVIE_ID
-import woowacourse.movie.view.detail.ReservationDetailActivity.Companion.TICKET
 import woowacourse.movie.view.home.ReservationHomeActivity.Companion.MOVIE_ID
+import woowacourse.movie.view.reservation.ReservationDetailActivity.Companion.DEFAULT_MOVIE_ID
+import woowacourse.movie.view.reservation.ReservationDetailActivity.Companion.TICKET
 import java.io.Serializable
 import java.text.DecimalFormat
